+++ js/lib/reject.js
@@ -1,28 +1,31 @@
 'use strict'
 
-const backCookie = require('../lib/createBackCookie').default
+const backCookie = require('../lib/createBackCookie')
 
-exports.default = ({publicCancelLoginUri, res, req, methods}) => {
-	var text = 'Unauthorized (401)'
-	res.status(401)
-	if(methods.basicAuth) {
+module.exports = ({publicCancelLoginUri, res, req, method, code}) => {
+	var text = 'Unauthorized'
+	if(method === 'basicAuth') {
+		res.status(401)
 		res.append(
 			'WWW-Authenticate',
 			'Basic Realm="Pls cancel this dialog if you forgot your password."'
 		)
+	}else{
+		res.status(200)
 	}
+
 	if(req) {
-		res.setHeader('Set-Cookie',backCookie({uri: req.originalUrl}))
+		res.setHeader('Set-Cookie', backCookie({uri: req.originalUrl}))
 	}
 	res.format({
 		'text/html': () => res.send(`<!DOCTYPE html>
 <html><body>
-	${text}
+	${text} (${code})
 	<script>document.location.href='${publicCancelLoginUri}'</script>
 </body></html>`
 		),
 		'application/json': () =>
-			res.send({status:{code:'ERROR', text}, publicCancelLoginUri}),
+			res.send({status:{code, text}, publicCancelLoginUri}),
 		default: () => res.send(text + '; see ' + publicCancelLoginUri)
 	})
 }