+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 1.3.0 2019-10-13
+* fix loginExpired to not redirect and update expiry on login page
+* add readme
+* increase test coverage to 100%
+
 ## 1.0.0 2019-01-31
 * initial version